Don't fail when participation already purchased
At the same time, also fix wrong data being part of event notifications.

+++ robozonky-app/src/main/java/com/github/triceo/robozonky/app/investing/Investor.java
@@ -114,7 +114,7 @@ public class Investor {
         final Investment i = Investor.convertToInvestment(r);
         this.zonky.invest(i);
         Investor.LOGGER.debug("Investment succeeded.");
-        return new ZonkyResponse(i.getAmount());
+        return new ZonkyResponse(i.getAmount().intValue());
     }
 
     private ZonkyResponse investOrDelegateOnCaptcha(final RecommendedLoan r) {
