+++ robozonky-app/src/test/java/com/github/triceo/robozonky/app/investing/StaticInvestorTest.java
@@ -28,6 +28,6 @@ public class StaticInvestorTest extends AbstractInvestingTest {
         final LoanDescriptor ld = AbstractInvestingTest.mockLoanDescriptor();
         final int recommended = 200;
         final Investment i = Investor.convertToInvestment(ld.recommend(recommended).get());
-        Assertions.assertThat(i.getAmount()).isEqualTo(recommended);
+        Assertions.assertThat(i.getAmount().intValue()).isEqualTo(recommended);
     }
 }
